requirements.txt
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi
gunicorn
pydantic
uvicorn
starlette
itsdangerous
